requirements.txt
# requirements.txt
py-cord
python-dotenv
requests
aiohttp